requirements.txt
torch==2.1.0
torch-npu==2.1.0.post12
torchaudio==0.13.0
torchmetrics==1.5.0
torchvision==0.16.0
numpy==1.23.5
av2==0.3.5
tqdm==4.67.1
tensorboard==2.20.0
hydra-core==1.3.2
lightning-fabric==2.4.0
lightning-utilities==0.11.9
pytorch-lightning==2.4.0
wandb==0.23.0
plotly==6.5.0
timm==1.0.22
safetensors==0.7.0
transformers==4.53.2
decorator==5.2.1
einops==0.8.1
ipdb==0.13.13
psutil==7.1.3











